Refactoring: rename ViteReact to React
