+++ src/main/java/tech/jhipster/lite/generator/client/react/core/domain/ReactService.java
@@ -1,0 +1,8 @@
+package tech.jhipster.lite.generator.client.react.core.domain;
+
+import tech.jhipster.lite.generator.project.domain.Project;
+
+public interface ReactService {
+  void addReact(Project project);
+  void addStyledReact(Project project);
+}
