+++ src/main/java/tech/jhipster/lite/generator/client/react/core/package-info.java
@@ -1,0 +1,2 @@
+@tech.jhipster.lite.BusinessContext
+package tech.jhipster.lite.generator.client.react.core;
