--- src/main/java/tech/jhipster/lite/generator/client/vite/react/core/application/ViteReactApplicationService.java
@@ -1,23 +1,0 @@
-package tech.jhipster.lite.generator.client.vite.react.core.application;
-
-import org.springframework.stereotype.Service;
-import tech.jhipster.lite.generator.client.vite.react.core.domain.ViteReactService;
-import tech.jhipster.lite.generator.project.domain.Project;
-
-@Service
-public class ViteReactApplicationService {
-
-  private final ViteReactService viteReactService;
-
-  public ViteReactApplicationService(ViteReactService viteReactService) {
-    this.viteReactService = viteReactService;
-  }
-
-  public void addViteReact(Project project) {
-    viteReactService.addViteReact(project);
-  }
-
-  public void addStyledViteReact(Project project) {
-    viteReactService.addStyledViteReact(project);
-  }
-}
