--- src/main/java/tech/jhipster/lite/generator/client/vite/react/core/domain/ViteReactService.java
@@ -1,8 +1,0 @@
-package tech.jhipster.lite.generator.client.vite.react.core.domain;
-
-import tech.jhipster.lite.generator.project.domain.Project;
-
-public interface ViteReactService {
-  void addViteReact(Project project);
-  void addStyledViteReact(Project project);
-}
