--- src/main/java/tech/jhipster/lite/generator/client/vite/react/core/package-info.java
@@ -1,2 +1,0 @@
-@tech.jhipster.lite.BusinessContext
-package tech.jhipster.lite.generator.client.vite.react.core;
