+++ src/main/java/tech/jhipster/lite/generator/packagemanager/npm/domain/NpmService.java
@@ -13,5 +13,5 @@ public interface NpmService {
 
   Optional<String> getVersion(String folder, String name);
   Optional<String> getVersionInCommon(String name);
-  Optional<String> getVersionInViteReact(String name);
+  Optional<String> getVersionInReact(String name);
 }
