+++ src/main/resources/generator/client/react/jest.config.ts
