+++ src/main/resources/generator/client/react/src/main/webapp/app/common/primary/app/App.tsx.mustache
