+++ src/main/resources/generator/client/react/src/main/webapp/app/index.css.mustache
