+++ src/main/resources/generator/client/react/src/main/webapp/app/vite-env.d.ts.mustache
