+++ src/main/resources/generator/client/react/src/main/webapp/config/setupTests.ts.mustache
