+++ src/main/resources/generator/client/react/src/main/webapp/content/images/JHipster-Lite-neon-blue.png
