+++ src/main/resources/generator/client/react/src/main/webapp/content/images/ReactLogo.png
