+++ src/main/resources/generator/client/react/src/main/webapp/index.html.mustache
