+++ src/main/resources/generator/client/react/src/test/javascript/spec/common/primary/app/App.test.tsx.mustache
