+++ src/main/resources/generator/client/react/tsconfig.json
