+++ src/main/resources/generator/client/react/vite.config.ts
