--- src/main/resources/generator/dependencies/vite/react/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "vite-react",
-  "version": "0.0.0",
-  "description": "JHipster Lite : used for Vite+React dependencies",
-  "license": "Apache-2.0",
-  "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "5.16.2",
-    "@testing-library/react": "12.1.3",
-    "@testing-library/user-event": "13.5.0",
-    "@types/jest": "27.4.1",
-    "@types/node": "17.0.21",
-    "@types/react": "17.0.39",
-    "@types/react-dom": "17.0.13",
-    "@vitejs/plugin-react": "^1.0.7",
-    "jest": "27.5.1",
-    "jest-css-modules": "2.1.0",
-    "jest-sonar-reporter": "2.0.0",
-    "react-scripts": "5.0.0",
-    "ts-jest": "27.1.3",
-    "ts-node": "10.6.0",
-    "typescript": "4.6.2",
-    "vite": "2.8.6"
-  }
-}
