--- src/test/java/tech/jhipster/lite/generator/client/vite/react/core/infrastructure/primary/rest/ViteReactResourceIT.java
@@ -1,67 +1,0 @@
-package tech.jhipster.lite.generator.client.vite.react.core.infrastructure.primary.rest;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static tech.jhipster.lite.TestUtils.readFileToObject;
-import static tech.jhipster.lite.common.domain.FileUtils.tmpDirForTest;
-import static tech.jhipster.lite.generator.client.vite.react.core.application.ViteReactAssert.assertAppWithCss;
-import static tech.jhipster.lite.generator.client.vite.react.core.application.ViteReactAssert.assertAppWithoutCss;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
-import tech.jhipster.lite.IntegrationTest;
-import tech.jhipster.lite.TestUtils;
-import tech.jhipster.lite.generator.client.vite.react.core.application.ViteReactAssert;
-import tech.jhipster.lite.generator.init.application.InitApplicationService;
-import tech.jhipster.lite.generator.project.domain.Project;
-import tech.jhipster.lite.generator.project.infrastructure.primary.dto.ProjectDTO;
-
-@IntegrationTest
-@AutoConfigureMockMvc
-class ViteReactResourceIT {
-
-  @Autowired
-  MockMvc mockMvc;
-
-  @Autowired
-  InitApplicationService initApplicationService;
-
-  @Test
-  void shouldInit() throws Exception {
-    ProjectDTO projectDTO = readFileToObject("json/chips.json", ProjectDTO.class).folder(tmpDirForTest());
-    Project project = ProjectDTO.toProject(projectDTO);
-    initApplicationService.init(project);
-
-    mockMvc
-      .perform(post("/api/vite/react").contentType(MediaType.APPLICATION_JSON).content(TestUtils.convertObjectToJsonBytes(projectDTO)))
-      .andExpect(status().isOk());
-
-    ViteReactAssert.assertDependency(project);
-    ViteReactAssert.assertScripts(project);
-    ViteReactAssert.assertReactFiles(project);
-    ViteReactAssert.assertFiles(project);
-    assertAppWithoutCss(project);
-  }
-
-  @Test
-  void shouldInitStyled() throws Exception {
-    ProjectDTO projectDTO = readFileToObject("json/chips.json", ProjectDTO.class).folder(tmpDirForTest());
-    Project project = ProjectDTO.toProject(projectDTO);
-    initApplicationService.init(project);
-
-    mockMvc
-      .perform(
-        post("/api/vite/react/styled").contentType(MediaType.APPLICATION_JSON).content(TestUtils.convertObjectToJsonBytes(projectDTO))
-      )
-      .andExpect(status().isOk());
-
-    ViteReactAssert.assertDependency(project);
-    ViteReactAssert.assertScripts(project);
-    ViteReactAssert.assertReactFiles(project);
-    ViteReactAssert.assertFiles(project);
-    assertAppWithCss(project);
-  }
-}
