+++ tests-ci/generate.sh
@@ -176,7 +176,7 @@ elif [[ $application == 'reactapp' ]]; then
   sonar_back_front
 
   callApi "/api/frontend-maven-plugin"
-  callApi "/api/vite/react/styled"
+  callApi "/api/react/styled"
 
 elif [[ $application == 'vueapp' ]]; then
   springboot
